[FEAT] Add error handling for delete button function

+++ public/js/detail.js
@@ -1,21 +1,23 @@
+const authority = document.querySelector('#authority').dataset.authority
 const deleteBtn = document.querySelector('.delete-btn')
-deleteBtn.addEventListener('click', function () {
-    axios({
-        method: 'delete',
-        url: `/delete/post/${location.pathname.match(/\/detail\/(.+)/)[1]}`
-    }).then(res => {
-        console.log(res)
-    }).catch(e => {
-        console.log(e)
-        const status = e.response.status
-        if (status === 401) {
-            const modal = document.querySelector('.modal')
-            modal.style.display = 'block'
-            setTimeout(() => {
-                modal.style.opacity = '1'
-            }, 10)
-        } else if (status === 403) {
-            alert(e.response.data)
-        }
+const editBtn = document.querySelector('.edit-btn')
+
+if (authority === "allowed") {
+    deleteBtn.addEventListener('click', async function () {
+        await axios({
+            method: 'delete',
+            url: `/delete/post/${location.pathname.match(/\/detail\/(.+)/)[1]}`
+        }).then(res => {
+            alert(res.data.msg)
+            location.href = res.data.url
+        }).catch(e => {
+            const { msg } = e.response.data
+            alert(msg)
+            location.reload(true)
+        })
     })
-})
+} else {
+    editBtn.style.display = 'none'
+    deleteBtn.style.display = 'none'
+}
+
